@@ -7,8 +7,9 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "antd": "^5.23.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "framer-motion": "^12.4.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.0",
     "styled-components": "^5.3.3",
